Added the ControllerAdvice to handle the common exception globally. Added unit tests, moved the handleNumberFormatException from RecipeController to ControllerAdvice.
